Fix TxBuilder generic errors.
This is an attempt to fix the tanglement of TxBuilder with Wallet's
generic.

`LocalUtxo`, `WeightedUtxo` and `Utxo` are made non generic. They should
not depend on keychain.

Some refactors of functions into generic implementation so they can be used
in the generic TxBuilder.

The last standing error is on `TxBuilder::finish_tx()` which calls
`wallet.create_tx()` which is only defined for the default wallet. To fix
this we need to change create_tx into a generic implementation.